refactor(SqlSchemaUpdater): Extract schema checks and upgrades to SqlSchemaUpdater.
