+++ src/main/java/com/conveyal/datatools/manager/DataManager.java
@@ -28,7 +28,7 @@ import com.conveyal.datatools.manager.extensions.transitland.TransitLandFeedReso
 import com.conveyal.datatools.manager.jobs.FeedUpdater;
 import com.conveyal.datatools.manager.persistence.Persistence;
 import com.conveyal.datatools.manager.utils.ErrorUtils;
-import com.conveyal.datatools.manager.utils.GtfsUtils;
+import com.conveyal.datatools.manager.utils.SqlSchemaUpdater;
 import com.conveyal.datatools.manager.utils.json.JsonUtil;
 import com.conveyal.gtfs.GTFS;
 import com.conveyal.gtfs.GraphQLController;
@@ -49,6 +49,8 @@ import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
+import java.sql.Connection;
+import java.sql.SQLException;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Properties;
@@ -116,10 +118,17 @@ public class DataManager {
         LOG.info("Data Tools server start up completed in {} seconds.", startupSeconds);
     }
 
+    /**
+     * Check that tables from namespaces referenced from projects/feed sources
+     * have the columns we need.
+     */
     private static void checkTables() {
-        // Check that tables from namespaces referenced from projects/feed sources
-        // have the columns we need.
-        GtfsUtils.checkReferencedNamespaces();
+        try (Connection connection = GTFS_DATA_SOURCE.getConnection()) {
+            SqlSchemaUpdater schemaUpdater = new SqlSchemaUpdater(connection);
+            schemaUpdater.checkReferencedNamespaces();
+        } catch (SQLException sqle) {
+            sqle.printStackTrace();
+        }
     }
 
     static void initializeApplication(String[] args) throws IOException {
