+++ src/test/java/com/conveyal/datatools/manager/TableUpdaterTest.java
@@ -1,7 +1,7 @@
 package com.conveyal.datatools.manager;
 
 import com.conveyal.datatools.DatatoolsTest;
-import com.conveyal.datatools.manager.utils.GtfsUtils;
+import com.conveyal.datatools.manager.utils.SqlSchemaUpdater;
 import com.conveyal.gtfs.loader.Table;
 import com.google.common.collect.Lists;
 import org.junit.jupiter.api.AfterAll;
@@ -31,7 +31,7 @@ public class TableUpdaterTest extends DatatoolsTest {
     @Test
     void shouldDetectMissingTableInNamespace() {
         List<String> excludedTables = Lists.newArrayList(Table.ROUTES.name, Table.TRIPS.name);
-        GtfsUtils.NamespaceInfo nsInfo = new GtfsUtils.NamespaceInfo("namespace", excludedTables);
+        SqlSchemaUpdater.NamespaceInfo nsInfo = new SqlSchemaUpdater.NamespaceInfo("namespace", excludedTables);
         assertEquals(
             String.join(",", excludedTables),
             nsInfo.missingTables.stream().map(t -> t.name).collect(Collectors.joining(","))
@@ -41,13 +41,13 @@ public class TableUpdaterTest extends DatatoolsTest {
     @Test
     void shouldDetectMissingTableColumns() {
         List<String> removedColumns = Lists.newArrayList("route_id", "route_short_name");
-        List<GtfsUtils.ColumnInfo> columns = Arrays
+        List<SqlSchemaUpdater.ColumnInfo> columns = Arrays
             .stream(Table.ROUTES.fields)
             .filter(f -> !removedColumns.contains(f.name))
-            .map(GtfsUtils.ColumnInfo::new)
+            .map(SqlSchemaUpdater.ColumnInfo::new)
             .collect(Collectors.toList());
 
-        GtfsUtils.TableInfo tableInfo = new GtfsUtils.TableInfo(Table.ROUTES, columns);
+        SqlSchemaUpdater.TableInfo tableInfo = new SqlSchemaUpdater.TableInfo(Table.ROUTES, columns);
         assertEquals(
             String.join(",", removedColumns),
             tableInfo.missingColumns.stream().map(c -> c.columnName).collect(Collectors.joining(","))
@@ -56,30 +56,30 @@ public class TableUpdaterTest extends DatatoolsTest {
 
     @Test
     void shouldDetectColumnsWithWrongType() {
-        List<GtfsUtils.ColumnInfo> columns = Arrays
+        List<SqlSchemaUpdater.ColumnInfo> columns = Arrays
             .stream(Table.ROUTES.fields)
-            .map(GtfsUtils.ColumnInfo::new)
+            .map(SqlSchemaUpdater.ColumnInfo::new)
             .collect(Collectors.toList());
 
-        GtfsUtils.TableInfo tableInfo = new GtfsUtils.TableInfo(Table.ROUTES, columns);
+        SqlSchemaUpdater.TableInfo tableInfo = new SqlSchemaUpdater.TableInfo(Table.ROUTES, columns);
         assertTrue(tableInfo.columnsWithWrongType.isEmpty());
 
         // Modify the type of one column
-        for (GtfsUtils.ColumnInfo c : columns) {
+        for (SqlSchemaUpdater.ColumnInfo c : columns) {
             if (c.columnName.equals("route_short_name")) {
                 c.dataType = "smallint";
                 break;
             }
         }
 
-        tableInfo = new GtfsUtils.TableInfo(Table.ROUTES, columns);
+        tableInfo = new SqlSchemaUpdater.TableInfo(Table.ROUTES, columns);
         assertEquals(1, tableInfo.columnsWithWrongType.size());
     }
 
     @Test
     void shouldDetectOrphanNamespace() {
         List<String> excludedTables = Arrays.stream(Table.tablesInOrder).map(t -> t.name).collect(Collectors.toList());
-        GtfsUtils.NamespaceInfo nsInfo = new GtfsUtils.NamespaceInfo("namespace", excludedTables);
+        SqlSchemaUpdater.NamespaceInfo nsInfo = new SqlSchemaUpdater.NamespaceInfo("namespace", excludedTables);
         assertTrue(nsInfo.isOrphan());
     }
 }
