+++ src/test/java/com/conveyal/datatools/manager/TableUpdaterTestE2E.java
@@ -7,7 +7,7 @@ import com.conveyal.datatools.manager.models.FeedSource;
 import com.conveyal.datatools.manager.models.FeedVersion;
 import com.conveyal.datatools.manager.models.Project;
 import com.conveyal.datatools.manager.persistence.Persistence;
-import com.conveyal.datatools.manager.utils.GtfsUtils;
+import com.conveyal.datatools.manager.utils.SqlSchemaUpdater;
 import com.conveyal.gtfs.loader.Table;
 import com.google.common.collect.Lists;
 import org.junit.jupiter.api.AfterAll;
@@ -17,7 +17,6 @@ import org.junit.jupiter.api.Test;
 import java.io.IOException;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
-import java.sql.SQLException;
 import java.util.ArrayList;
 
 import static com.conveyal.datatools.TestUtils.appendDate;
@@ -68,20 +67,22 @@ public class TableUpdaterTestE2E extends UnitTest {
      * and to upgrade the tables accordingly.
      */
     @Test
-    void canCheckAndUpgradeTables() throws SQLException {
+    void canCheckAndUpgradeTables() {
         // Create source version (folder contains stop_attributes file).
         FeedVersion sourceVersion = createFeedVersionFromGtfsZip(
             feedSource,
             "caltrain_gtfs_lite.zip"
         );
 
-        try (Connection connection = DataManager.GTFS_DATA_SOURCE.getConnection()) {
+        try (
+            Connection connection = DataManager.GTFS_DATA_SOURCE.getConnection();
+        ) {
+            SqlSchemaUpdater schemaUpdater = new SqlSchemaUpdater(connection);
             String namespace = sourceVersion.namespace;
-            GtfsUtils.NamespaceInfo nsInfo = GtfsUtils.checkTablesForNamespace(
+            SqlSchemaUpdater.NamespaceInfo nsInfo = schemaUpdater.checkTablesForNamespace(
                 namespace,
                 "test version",
-                "namespace",
-                connection
+                "namespace"
             );
 
             // Some tables are missing from the feed and should be flagged.
@@ -94,7 +95,7 @@ public class TableUpdaterTestE2E extends UnitTest {
                 Table.TRANSLATIONS
             )));
             for (Table t : nsInfo.validTables) {
-                GtfsUtils.TableInfo tableInfo = GtfsUtils.checkTableColumns(nsInfo, connection, t);
+                SqlSchemaUpdater.TableInfo tableInfo = schemaUpdater.checkTableColumns(nsInfo, t);
                 assertNotNull(tableInfo);
 
                 // The agency table is missing columns, so they should be flagged.
@@ -108,7 +109,7 @@ public class TableUpdaterTestE2E extends UnitTest {
                     );
 
                     assertEquals(missingColumns.size(), tableInfo.missingColumns.size());
-                    for (GtfsUtils.ColumnInfo c : tableInfo.missingColumns) {
+                    for (SqlSchemaUpdater.ColumnInfo c : tableInfo.missingColumns) {
                         assertTrue(missingColumns.contains(c.columnName));
                         assertEquals(
                             String.format(
@@ -132,11 +133,11 @@ public class TableUpdaterTestE2E extends UnitTest {
                     changeStatement.execute();
 
                     // Check that the modified column is flagged.
-                    GtfsUtils.TableInfo changedTableInfo = GtfsUtils.checkTableColumns(nsInfo, connection, t);
+                    SqlSchemaUpdater.TableInfo changedTableInfo = schemaUpdater.checkTableColumns(nsInfo, t);
                     assertNotNull(changedTableInfo);
 
                     assertEquals(1, changedTableInfo.columnsWithWrongType.size());
-                    GtfsUtils.ColumnInfo columnWithWrongType = changedTableInfo.columnsWithWrongType.get(0);
+                    SqlSchemaUpdater.ColumnInfo columnWithWrongType = changedTableInfo.columnsWithWrongType.get(0);
                     assertEquals("monday", columnWithWrongType.columnName);
                     assertEquals("varchar", columnWithWrongType.dataType);
                     assertEquals("integer", columnWithWrongType.expectedType);
@@ -145,19 +146,18 @@ public class TableUpdaterTestE2E extends UnitTest {
             }
 
             // Go ahead and update the tables.
-            GtfsUtils.upgradeNamespace(nsInfo, connection);
+            schemaUpdater.upgradeNamespace(nsInfo);
 
-            GtfsUtils.resetScannedNamespaces();
-            GtfsUtils.NamespaceInfo updatedNsInfo = GtfsUtils.checkTablesForNamespace(
+            schemaUpdater.resetScannedNamespaces();
+            SqlSchemaUpdater.NamespaceInfo updatedNsInfo = schemaUpdater.checkTablesForNamespace(
                 namespace,
                 "test version",
-                "namespace",
-                connection
+                "namespace"
             );
 
             // Check that missing fields were added
             for (Table t : updatedNsInfo.validTables) {
-                GtfsUtils.TableInfo tableInfo = GtfsUtils.checkTableColumns(nsInfo, connection, t);
+                SqlSchemaUpdater.TableInfo tableInfo = schemaUpdater.checkTableColumns(nsInfo, t);
                 assertNotNull(tableInfo);
 
                 // The agency table is missing columns, so they should be flagged.
@@ -175,12 +175,14 @@ public class TableUpdaterTestE2E extends UnitTest {
             // Check that missing tables were added.
             assertTrue(updatedNsInfo.missingTables.isEmpty());
             for (Table t : updatedNsInfo.validTables) {
-                GtfsUtils.TableInfo tableInfo = GtfsUtils.checkTableColumns(updatedNsInfo, connection, t);
+                SqlSchemaUpdater.TableInfo tableInfo = schemaUpdater.checkTableColumns(updatedNsInfo, t);
                 assertNotNull(tableInfo);
 
                 assertTrue(tableInfo.missingColumns.isEmpty());
                 assertTrue(tableInfo.columnsWithWrongType.isEmpty());
             }
+        } catch (Exception e) {
+            e.printStackTrace();
         }
     }
 }
